New upstream version 1.0.1

+++ virtinst/urlfetcher.py
@@ -63,8 +63,15 @@ class _ImageFetcher(object):
     def saveTemp(self, fileobj, prefix):
         if not os.path.exists(self.scratchdir):
             os.makedirs(self.scratchdir, 0750)
-        (fd, fn) = tempfile.mkstemp(prefix="virtinst-" + prefix,
-                                    dir=self.scratchdir)
+
+        prefix = "virtinst-" + prefix
+        if "VIRTINST_TEST_SUITE" in os.environ:
+            fn = os.path.join(".", prefix)
+            fd = os.open(fn, os.O_RDWR | os.O_CREAT)
+        else:
+            (fd, fn) = tempfile.mkstemp(prefix=prefix,
+                                        dir=self.scratchdir)
+
         block_size = 16384
         try:
             while 1:
@@ -182,10 +189,16 @@ class _MountedImageFetcher(_LocalImageFetcher):
     Fetcher capable of extracting files from a NFS server
     or loopback mounted file, or local CDROM device
     """
+    _in_test_suite = bool("VIRTINST_TEST_SUITE" in os.environ)
+
     def prepareLocation(self):
         cmd = None
-        self.srcdir = tempfile.mkdtemp(prefix="virtinstmnt.",
-                                       dir=self.scratchdir)
+
+        if self._in_test_suite:
+            self.srcdir = os.environ["VIRTINST_TEST_URL_DIR"]
+        else:
+            self.srcdir = tempfile.mkdtemp(prefix="virtinstmnt.",
+                                           dir=self.scratchdir)
         mountcmd = "/bin/mount"
 
         logging.debug("Preparing mount at " + self.srcdir)
@@ -200,21 +213,24 @@ class _MountedImageFetcher(_LocalImageFetcher):
 
         logging.debug("mount cmd: %s", cmd)
 
-        ret = subprocess.call(cmd)
-        if ret != 0:
-            self.cleanupLocation()
-            raise ValueError(_("Mounting location '%s' failed") %
-                             (self.location))
+        if not self._in_test_suite:
+            ret = subprocess.call(cmd)
+            if ret != 0:
+                self.cleanupLocation()
+                raise ValueError(_("Mounting location '%s' failed") %
+                                 (self.location))
         return True
 
     def cleanupLocation(self):
         logging.debug("Cleaning up mount at " + self.srcdir)
-        cmd = ["/bin/umount", self.srcdir]
-        subprocess.call(cmd)
-        try:
-            os.rmdir(self.srcdir)
-        except:
-            pass
+
+        if not self._in_test_suite:
+            cmd = ["/bin/umount", self.srcdir]
+            subprocess.call(cmd)
+            try:
+                os.rmdir(self.srcdir)
+            except:
+                pass
 
 
 class _DirectImageFetcher(_LocalImageFetcher):
